gpu: raytrace_to_texture: Add missing exposure.
This would be much simpler if we did exposure in the postprocessing
pass which everything shares, but that would require guaranteed f16
support which we're trying not to require, and either a g-buffer
identifying which layer each pixel belongs to or a separate
postprocessing pass for each layer.

This commit also introduces `ColorBuf::into_premultiplied_rgba()`,
which should probably now be used more than it is.