gpu: raytrace_to_texture: Refine rays_per_frame updating.
Besides adding a finer-grained adjustment policy, the use of the
framebuffer area was incorrect and left over from before `PixelPicker`
was as fancy as it is. Now we use `cycle_length()` instead, which
may or may not be equal to the framebuffer area.

(`rays_per_frame` is a misnomer given background updating, and it
ought to be renamed to `rays_per_batch` or something like that.)