raycast: Introduce `Parameters` struct.
This is a pure refactoring which I hope will enable future improvements
in actual functionality, as described in `Parameters`’ docs.

According to benchmarks, it seems to cause a slight increase in time,
but that seems to be sensitivity to struct field ordering, which I don’t
care to try to optimize any time soon (and probably depends on
architecture and what the step code gets inlined into, anyway).