space: Add benchmark for `LightUpdateQueue`.
`LightUpdateQueue` is a substantial part of the cost of light updates,
so we should examine it in more isolation, with this benchmark.

This commit also moves the rest of the "public for testing" items in
`all_is_cubes::space::light` over to the `_special_testing` mechanism
(which has the benefit of dead code analysis when the items are *not*
being exposed for testing).