space: Reimplement `LightUpdateQueue` on top of `priority_queue` library and remove `fallback_priority`.
This greatly speeds up `fast_evaluate_light()`, which spends most of its
time adding entries to the queue, and is neutral overall in the lighting
evaluation benchmarks.

Therefore, it will reduce the maximum possible hang from initializing
light in a space, which is worth doing.

We lose the clever fallback priority, but get some happenstance behavior
that’s acceptable (and deterministic since the seed is not random).