mesh: Redefine `DepthOrdering` with 27 cases instead of 48.
The “volumesort” approach’s 48 cases:

* Distinguish between different diagonals that we don’t need.
* Don’t distinguish between “within-bounds” cases that we do need.

The new scheme uses a 3×3×3 cube of cases, where each axis has the cases
where the viewpoint is lower than the mesh, within the mesh’s projection
on that axis, or higher than the mesh. When the viewpoint is either
lower or higher on all 3 axes, this is equivalent to the old scheme.
When the viewpoint is within on 1 or 2 axes, it produces different
incorrect results than the old scheme did.

Next steps:
* When within on 2 axes, within those tight bounds, perform depth
  sorting. (This means that we will have more than 1 depth sorted chunk.
  Such is the price we pay for having actually correct results.)
  This should fix the incorrect drawing bug.
* Pick the `DepthOrdering` based on the actual mesh bounding box rather
  than the abstract chunk bounds ignoring the content. This reduces
  the amount of sorting needed.

See <https://github.com/kpreid/all-is-cubes/issues/53> for history.