mesh: Add `IndexSliceMut`.
This is a natural extension which previously didn’t have any specific
use, but now it has one in `dynamic_depth_sort_for_view()`.
In particular, future work on the depth sorting bug may involve, as an
optimization, sorting multiple slices in parallel (where we know they
are already sorted relative to each other), at which point we will want
a `split_at_mut()` for `IndexSliceMut` or something similar to provide
disjoint access.