mesh: Remember a region of validity of depth sorts.
Sorting is skipped if the view position is still within the region.

I’m adding this so that I can do sorting of more chunks (to fix the
depth ordering bug <https://github.com/kpreid/all-is-cubes/issues/53>)
without greatly increasing the CPU cost.

I’m not sure this strategy is completely correct; in fact, I think it
might technically require that we change the sort order to use Manhattan
distance instead of Euclidean distance. But, that will rarely matter.