mesh: Require `DynamicMeshTypes` vertex coordinates to be `Send + Sync`.
This will be needed for upcoming depth sorting changes to run the sorts
in parallel along with chunk updates. It’s necessary to write this bound
explicitly because, technically, the vertex type is not actually
required to contain the coordinate type; or, there could be an
`unsafe impl Send for MyVertex` allowing a vertex type containing an
non-`Send` coordinate type, not that that makes much sense.