mesh: Switch to `f32` mesh coordinates and remove `Vertex::Coordinate`.
Since `BlockMesh` coordinates range from 0 to 1, and `SpaceMesh`
coordinates are relative to the low corner of the requested bounds
(as opposed to using the origin of the surrounding `Space`), we can
always use `f32` coordinates for vertex positions. The only fiddly bit
is that when we compare a bounding box to a camera position (for
culling and `DepthOrdering`), we want to do that by converting the
camera position to relative coordinates first.

The `Coordinate` associated type on `Vertex` is gone, but vertex types
are still allowed to have any format of coordinate storage they like,
as long as they can convert it back to `f32` on request.

Benefits:
* Smaller data structures.
* Constructing `SpaceMesh`es is up to 28% faster by benchmark.
* Improved build time and binary size (due to less generic code?)