render: Add `LightingOption::Coarse`.
The idea is that this hides the interpolation artifacts (e.g sharp
diamond-like shapes) from linear interpolation behind deliberate
blockiness. It might also work well with when we get better light in
volumetric blocks, allowing consistency while still taking a low number
of samples. However, for now it is just an option I am adding so that
we can experiment with how it looks in various cases.