mesh: Add `dynamic_sub_ranges`.
This field in `TransparentMeta` will be useful in the future to optimize
depth sorting by sorting smaller ranges separately. It also neatly
eliminates a bunch of individual checks for trivial cases in
`dynamic_depth_sort_for_view()`, because `dynamic_sub_ranges.is_empty()`
is true in all of them. (We leave in the test for `WANTS_DEPTH_SORTING`
because that makes that case statically visible to the optimizer.)