math: Simplify `Aab::wireframe_points()`.
The goal of this change is to reduce build time and binary size by
replacing a complex piece of code with a table. Oddly, it only half
achieves that goal: the size of the wasm binary goes down a bit
and the size of the aarch64 binary goes up a bit. Still, better that
than the opposite.