math: Implement `GridRotation` multiplication/composition using a lookup table.
This makes the machine code significantly less complex, which should be
good for performance. It seems to have a small positive effect on
`cargo bench --bench gen`, which is where we might expect to see the most
composed rotations.

The binary size for aarch64-apple-darwin is just +2 kB larger, slightly
less than the size of the table (48 × 48 × 1 B = 2.3 kB). The binary size
for wasm32-unknown-unknown is greatly reduced (-17 kB).

I haven’t got a positive or negative benchmark result.