space: Inline `Space::mutate()`.
This is for the purpose of code size reduction, not speed.
Because `mutate()` is always called with a unique closure, there is no
particular benefit to having out-of-line copies of the function; each
one would only be called once.

Ideally the construction of `Mutation` would be in a non-generic
function, but that would be messy, given the combination of the borrow
splitting, the temporary notification buffers, and the generic return
type.