Replace `tokio` and `futures-executor` with `smol` in tests and benches.
We don’t actually need the functionality of Tokio, and using a
simpler executor should reduce build time and binary size (though
not total workspace build time, because `tokio` is still in the
workspace and will continue to be due to `all-is-cubes-server`).

(`test-renderers` still uses Tokio because it makes use of Tokio’s
task-local variables. Perhaps I will change that later.)