desktop: Replace `tokio` executor with `smol`.
`all-is-cubes-desktop` is still using `tokio::sync` types.
Those will be replaced in the next commit.

We don’t actually need the functionality of Tokio, and using a
simpler executor reduces build time and binary size (though not total
workspace build time, because `tokio` is still in the workspace and
will continue to be due to `all-is-cubes-server`).