mesh: Remove obsolete `layer_is_visible_somewhere` logic.
This check is now done, better, in the separate analysis phase, so it
will always be `true` and we don’t need to calculate it.

For some reason, the mesh benchmark says this is an *increase* in
execution time even though this doesn’t even change much in the
assembly. But we can’t let benchmark weirdness stop us from deleting
code.