math: Add `Rgb01` type.
This color type restricts the RGB components to strictly [0, 1], and
thus expects that they are relative (e.g. proportion of light reflected,
or a SDR image pixel) rather than any kind of absolute light intensity
as `Rgb` is used for.

This is part of a project to refine our color types so that their
meaning is clearer and it is harder to make mistakes. For context, see
<https://github.com/kpreid/all-is-cubes/issues/679>.