mesh: Fix bug with coplanar transparent and opaque voxels.
In `TransparencyFormat::BoundingBox`, the *only* transparent parts of
the mesh should be the singular bounding box, but instead, if the
bounding box of the opaque voxels in a single plane included any
transparent voxels, the compute phase would include them as if in
`TransparencyFormat::Surfaces` mode, resulting in rendering that
transparent volume twice when viewed through that region of the plane.

This change fixes that bug by suppressing all transparent voxels in
`get_voxel_with_limit()` during the compute phase, whenever using
`BoundingBox` format.

Note: This change introduces a discrepancy between mesh mand raytraced
rendering in the `TRANSPARENCY_GLASS_AND_WATER` exhibit. This is a
coincidence, because the raytracer has an unrelated overcounting bug