raycast: Optionally include steps which exit the bounds.
This information will help the raytracer's thickness calculations by
making it possible to compute the intersection point of the exit,
but that change is not included in this commit.

It is arguably more consistent in general, as long as we interpret
`RaycastStep` as always representing the crossing between two cubes.
However, multiple use cases only care about reaching in-bounds cubes,
so we also still offer the mode where the exit step is not produced.