raytracer: Account for exiting voxel data bounds.
This fixes a bug where if a block is transparent and smaller than its
cube on the *back* side,

    . . . +-+ . . . . .
    .     | |   .     .
    ——————|—|——>.     .
    .     | |   .     .
    . . . +-+ . . . . .
    0     1     2     3

the raytracer would incorrectly count the material thickness as
occupying 1.0 .. 2.0 rather than 1.0 .. 1.5.

This change works together with the preceding change where the raycaster
now can report steps that exit the given bounds.