port: Intern .vox palette entries.
The new `struct ExportPalette` keeps track of which palette entries
have been created, to avoid creating identical entries due to exporting
multiple models or because of differing properties that are ignored by
export.

There is also support for .vox material properties, such as light
emission and transparency, but this is disabled because `dot_vox`
currently writes them in a nondeterministic order, which breaks
our snapshot testing.