math: Tweak `Cube::containing()` code.
This was originally intended as an attempt to make `Cube::containing()`
a `const fn`, eliminating all uses of generics, but `floor()` is not
available in `no_std` in a const form.

However, this change produces shorter machine code and notable
binary size savings (13 kB desktop, 0.7 kB wasm), and no performance
effect noticeable in the raycast bench, so it seems worth keeping.
(I didn't check whether the `unlikely_none` actually helps.)