raytracer: Add `Exception` enum to consistently describe exceptional situations.
This replaces `RtBlockData::{error,sky}()` with a single `exception()`
function, and is also available as a field of `Hit`.

This allows `Accumulate` and `RtBlockData` to more consistently receive
information about what’s going on and react to it as they see fit.

I plan to use this mechanism to replace the current kludgy
implementation of `debug_pixel_cost`.