raytracer: Use a new `Exception` value to implement `debug_pixel_cost`.
This replaces the hidden special case `_unstable_get_original_color()`
with a new hidden special case `Exception::DebugOverrideRg`, but the new
one is better behaved; it works even when `ColorBuf` is in a wrapper.

More importantly, it avoids replacing the entire `Accumulator` value.
This is intended to allow us to stop requiring `Default`, and will
also allow other data in the accumulator to be preserved.