Replace `|_| {}` and `|x| x` closures with `drop` and `identity`, where possible.
This is possible only when the needed function signature doesn’t have
any lifetimes (so, not for `RenderDataUpdate`).

I was only hoping this would save a little compile time by not
introducing unnecessarily distinct function definitions, but it actually
reduces the binary size a bit — presumably due to fewer symbols, because
it isn’t a difference in machine code.

Note that this only actually reduces monomorphizations if the named
function is used *more than once*; but I did this replacement everywhere
for consistency.