universe: Move pending handle values to transactions.
Part of <https://github.com/kpreid/all-is-cubes/issues/633>.
In order for handles to be unconditionally `Sync`, they need to not
contain any `Mutex`; handle states must be managed purely by
lock-free atomics.

This doesn’t strictly prohibit keeping the value there, now that we’ve
prohibited mutation of values in pending handles, but I think that this
will be cleaner in the long run anyway. Also, it prevents pending
handles from forming strong reference cycles that leak memory if their
transaction is not committed.