transaction: Commit `Transaction`s by value instead of reference.
This should reduce the amount of cloning involved in the typical
transaction execution. In particular, most clones are now in tests.
If it turns out that there is a situation where we absolutely need to
keep a transaction value without cloning, we can handle that by
adding `impl Transaction for &Tr` for the relevant transaction type.

The actual motivation for this change, however, is to make it possible
to define transaction types that do not implement `Clone` and can be
committed only once, and potentially make `UniverseTransaction` be
such.