gpu: raytrace_to_texture: Fix layer detection for reprojection.
This was broken by 93146d58a58d7593fc4788c3372fc474ee0e5c2d which
started using a single accumlator for both layers. Eventually we want to
keep fully separate buffers for the layers, and the `InLayer` will be
moot, but that will be more complex than this simple fix.