desktop: Disable lighting in test end-to-end/output-path-with-graphics.
This means the test, which is intended to test *whether* the two outputs
are written at all and not the details of graphical output, will not
break every time a small lighting change occurs.

This is now necessary because commit 1705c8b6ac49020df8bd497a458a9dfd505ce13f
made the ordering of light updates platform-dependent; while that change
itself was unforeseen (I neglected to think about the fact that the
hash table’s hash function might be platform-dependent), I do actually
want to allow *nondeterminism* of light updates so as to be able to
use execution strategies that are unable to be deterministic (e.g.
writing individual cube values atomically from a background thread).