universe: Remove `impl Clone for UniverseTransaction`.
This simplifies `UniverseTransaction`’s job by not requiring it to be
able to be cloned meaningfully; in particular, it may uniquely own the
data of the members it would insert when committed (when we move that
data out of the `Handle` for <https://github.com/kpreid/all-is-cubes/issues/633>).

This commit does not do any of that, but only removes the impl.
The rest of the modified code is either tests that used cloning for
convenience, or which tested now-impossible situations.