math: Move `Vol::index()`’s algorithm to a non-generic function.
By itself, this is to improve build performance, but I also want to
experiment with adding variations like SIMD or unchecked indexing,
which suggest this will probably be broken up into more functions,
which should not live inside or next to the `impl`.

Also added section-dividing comments to the vol.rs file.