universe: UniverseTransaction: Use a keyed collection for `anonymous_insertions`.
This will be necessary to allow reading the stored values by handle,
once we move the values from the pending handles to the transaction.

It’s possible that in the long run, we won't actually need the ability
to read from transactions, in which case this change should be reverted
for efficiency.