universe: Remove unnecessary complication in `Handle::insert_pending_into_universe()`.
It used to be that pending handles were created independent of
transactions, and could be inserted by whoever got there first.
Now, only one transaction can possibly be doing the insert, so we can
be a lot less careful — in particular, not hold the state lock across
everything else. The next step in this direction is to replace the
state mutex completely with lock-free atomics.