Remove `maybe_sync` module and conditional compilation.
This is a step towards restoring `no_std` functionality.

Instead of making `Send + Sync` conditional, we require it always,
which is necessary for anything going into a `bevy_ecs` component,
and eliminates feature non-additivity.

The price of this is that we use `bevy_platform`'s mutex shim, which
uses `std::sync` on `std`, and spinlocks everywhere else, which is fine
as long as we are not running on a platform that is both `no_std` and
multithreaded. This is not ideal, but better than the broken status quo.

The remaining work is largely to find ways to make `nosy` work with
`Sync` + no_std.