gpu: Antialias texel edges.
This means that a flat surface of voxels won't have non-antialiased
texel edges in between MSAA antialiased block edges.

It seems like there might be something in this change causing texture
atlas bleeding, visible mostly in `--template=dungeon`, but it is rare
and I can’t consistently reproduce it enough to debug it.

Note that the render tests still have different images — this is not
because of differences in the edge antialiasing results themselves, but
rather the distant areas of the test scene which still exhibit aliasing
due to insufficient sampling. Those areas need mipmaps/LODs, not just
4xMSAA.