content: Add `framework` module for supporting procedural generation.
The capabilities of `framework::Queue` are currently quite limited;
the main thing it can do is replace a basic async yielding and
progress-reporting loop. Eventually it should gain abilities to
subdivide tasks and perform basic layout like box walls, and its
state and intermediate steps should be user-inspectable.

The goal of this change is to reduce the marginal complexity of each
new procedural generation feature. Right now, it doesn’t achieve that;
if the framework doesn't get significantly more powerful than it
currently is, it should probably be reverted.