mesh: Fix hole processing bug in new triangulator.
At least, I hope it’s fixed.

We skip a triangle if it would be inverted (which seems to be a reliable
indicator of trouble) and then later add the missing triangles using ear
clipping.

There could probably be some useful additional optimizations, but I
think this is a *robust* solution to the problem.