mesh: Reuse PlanarTriangulator for all faces.
This doesn't show up on the benchmarks like I was hoping, but 1/6th the
allocations has got to be good, right? It will also be useful later for
adding transparent surface support, because the plan I have for doing
that is to do separate passes for transparent and opaque surfaces that
have separate flags in the `PtBasis`.