Universe: Turn `mutate_character()` into `mutate_component()`.
This allows accessing any component as long as it implements the
dedicated trait `PubliclyMutableComponent`.

This removes the requirement that `Character` is a component. Further,
I think this will be a good solution in general for making it easy to
perform special-case accesses to the ECS that don't fit into the
transaction paradigm, such as player input to characters.