sound: Add skeleton of data structures for ambient sound.
Ambient sound is sound that is not emitted at a time by a simulated
event, but is derived from the listener’s immediate environment.

This commit adds a (nonfunctional) ambient sound attribute to blocks.
Future work will be to actually have data fields in `Ambient`,
gather the character’s neighborhood into a summary similar to exposure
calculation, and synthesize and play the corresponding audio.