character: Make `exposure::State` a separate ECS component.
Ambient sound will require a very similar calculation (that can’t
share much or any of the work because it is direction-sensitive in
a different way), so separating it out will allow those systems to
run in parallel.

Also simplified the code of `step_exposure()`, eliminating the only use
of `universe::ReadMember`.