sound: Remove `absorption_bands` and add `Spectrum` struct.
`absorption_bands` won’t be part of the first pass implementation of
the whole system, so we shouldn’t have a nonfunctional stub.

`Spectrum` will help us make things correct and ergonomic when more
complex things start happening. In particular, it can implement
operators in a well-defined way, which is a better plan than
implementing operators for all of `Ambient` with unclear meaning.