space: Wrap `BlockListener`s in `Arc` early.
This is refactoring in preparation for making the listener gate
integral to the listener, which is in preparation for being able to
create listeners before knowing their block index, which will enable
evaluating blocks during transaction check instead of commit.

Also, `SpaceBlockData::new` is now non-generic, because it had no actual
reason to be generic. (That could have been a separate change.)