space: Integrate listener gates into  `BlockListener` itself.
`Arc<()>` is the implementation of `nosy::Gate`, so this does exactly
the same thing, just integrated instead of as a generic wrapper.

This is refactoring in preparation for being able to create listeners
before knowing their block index, which will enable evaluating blocks
during transaction check instead of commit. Specifically, we will be
able to hang onto an `Arc<BlockListener>` (or `Weak<BlockListener>`)
for adding the block index later, without needing to add another `Arc`
to the system.