gpu: Use alpha testing in addition to texture coordinate clamping.
The goal of this is to stop requiring all-is-cubes-mesh to supply
per-triangle clamp boxes, which will allow more vertex sharing and
thus be more compatible with the new triangulator (as well as making
meshes smaller). We still need clamp boxes at the granularity of
whole blocks / atlas allocations for:

* Texels on the edge of the atlas allocation getting mixed with
  texels belonging to different allocations.
* Terminating raymarching.

Both of these could in principle be solved by padding in the atlas that
has a sentinel value, but for now, this is a smaller change.