gpu: Use clamp boxes from texture allocation instead of `Coloring`.
This will allow removing the clamp boxes from all-is-cubes-mesh,
unblocking the new triangulator. The previous commit changed shading
so that we do not need the more precise boxes we are getting rid of.

The texture coordinates are now `@interpolate(perspective, centroid)`
so that in all of the cases where we do not use the linear texture
sampling trick at all, but are still using MSAA, we get in-bounds
samples rather than possibly extrapolated ones. (If we weren’t doing
the linear sampling for antialiased texel edges within triangles,
then `centroid` would solve the entire problem and we wouldn’t need
any clamp boxes at all.)