mesh: Remove texture coordinate clamp boxes.
As of the previous commit, they are ignored by all dependents.

If they are desired, they are now the responsibility of the caller,
who can obtain the needed information from their own texture tiles.

This change is needed to unblock the new triangulator, which does not
naturally produce separate vertices per quad and thus would need a
separate vertex splitting step to produce tight clamp boxes.