mesh: Generalize depth sorting to non-rectangles.
This is necessary in order to be able to use the new triangulator for
all mesh construction including `TransparencyFormat::Surfaces`. Note
that this change does not seem fully correct — it does not panic, but
the depth sorting results are not visually correct.

In order to avoid a performance regression, we sort as rectangles
(previously referred to less precisely as quads) if the mesh in fact
contains only rectangles (which is always the case under
`TransparencyOption::Volumetric`).