mesh: Track narrower range of ear-clipping passes.
This should cut down on the worst-case time when the frontier has many
vertices not eligible for the ear clipping. In benchmarks it is neutral
to faintly worse, but I have reason to believe that's more a codegen
fluke than a meaningful result, because *just* adding the` if` around
the call to `clip_ears_in_new_frontier()`caused a worse increase in
benchmark times.

The external `if` and `#[cold]` is largely irrelevant to the algorithm
change *per se*; I did it mainly to get better profile results by not
inlining the function — but it *is* true that the ear clipping rarely
needs to be run, so not-inline and cold should be a good choice.