port: Better error handling in parsing .vox files.
Generally, this is replacing unwraps with error propagation.
Some of these cases probably can’t actually be hit (e.g. `u32` to
`usize` conversion is actually infallible on all platforms we support)
but “no unwraps on the input data at all” is a good principle to stick
to in parsing.