content: Shrink default size of the `CornellBox` template.
The point of the template is to exercise how the lighting algorithm
behaves on a standard computer graphics test case, but the previously
chosen size was “unrealistically” large for the kind of content I
intend to support, so it was setting an excessively high standard.

It still doesn’t look good at the new size, but fixing that is
in-scope for what I want the lighting algorithm to do.