raytracer: Draw info text with 8-neighbor outline instead of 4-neighbor.
I’ve decided that info text should be drawn in this style everywhere
— the shadows used on the GPU renderer are sometimes not very legible,
I am planning to rewrite text rendering in a way that will make using,
a blur more costly, and pixelated edges are more our style anyway.

This commit implements the new style for the raytracer and an upcoming
commit will implement it for the GPU renderer.