test-renderers: Expand known-incorrect-renderer mechanism.
This disables the antialiasing test on software renderers that are known
to do odd things with it. Eventually we should troubleshoot the
misbehavior to determine if it is actually our mistake, but this is
still an improvement for now, because it will let us run the other tests
on Linux and Windows software renderers.

Had to fix `rendiff` to correctly handle the new `Threshold` value used
here, hence the dep update. (I could have written the `Threshold`
differently, but the old behavior was simply a bug I had not fixed.)