block: Add `Modifier::SetAttribute`  replacing `Modifier::Attributes`.
`Modifier::Attributes` did not make sense as a modifier because it
replaces all attributes, rather than only the ones actually intended to
be set to specific values. For example, it was not possible to set the
name of a `Composite` block without overriding all other attributes.

`Modifier::SetAttribute` does not have this problem, and also creates
consistency in the user/developer content authoring experience, the
serialization format, and the in-memory storage of blocks — all three
contain only explicitly set attributes, and `BlockAttributes` is now
a product of block evaluation and never serialized.

Note that in the future, we might want to switch to custom modifiers
that set different groups of attributes than exactly one. The intended
direction here is not specifically the existence of `SetAttribute`, but
the lack of `Modifier::Attributes`.