block: Add internal `BlRotate` trait for rotatable block attributes.
This trait generalizes the pattern of offering `rotate()` and
`rotationally_symmetric()`, and allows those methods of
`BlockAttributes` to be implemented via macro.

This is of only small benefit now, but is intended to be useful when I
introduce modifiers that set individual attributes, which need to rotate
identically to how the attributes do in other contexts.