raytracer: Move code from `all-is-cubes` to `all-is-cubes-render`.
Only the parts shared with block evaluation remain in `all-is-cubes`.

This reduces the number of re-export shenanigans, should improve
compilation time by moving code out of the foundational `all-is-cubes`
crate, and brings the normal and orthographic raytracers into the same
crate.

In the future, we might also move `GraphicsOptions` the same way, since
it makes lots of assumptions about renderers.